refactor: home header component
