+++ frontend/src/components/features/home/home-header/guide-message.tsx
